Created "Priorities.txt"
This file will be in the repository as a way to keep track of issues and priorities via in the actual source rather than GitHub. Basically a failsafe. 

Also, it has a "Suggestions" area in it that we can add to, suggestions that can possibly be put in the game. 
*Do not remove.*